+++ Example/adapter/src/main/java/com/google/ads/mediation/sample/adapter/SampleNativeAppInstallAdMapper.java
@@ -25,6 +25,7 @@ import android.view.ViewParent;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 
+import com.google.ads.mediation.sample.sdk.SampleMediaView;
 import com.google.ads.mediation.sample.sdk.SampleNativeAppInstallAd;
 import com.google.android.gms.ads.formats.NativeAd;
 import com.google.android.gms.ads.formats.NativeAdOptions;
@@ -68,6 +69,16 @@ public class SampleNativeAppInstallAdMapper extends NativeAppInstallAdMapper {
         extras.putString(SampleAdapter.DEGREE_OF_AWESOMENESS, ad.getDegreeOfAwesomeness());
         this.setExtras(extras);
 
+        SampleMediaView mediaView = mSampleAd.getMediaView();
+
+        // Some ads from Sample SDK has video assets and some do not.
+        if (mediaView != null) {
+            setMediaView(mediaView);
+            setHasVideoContent(true);
+        } else {
+            setHasVideoContent(false);
+        }
+
         setOverrideClickHandling(false);
         setOverrideImpressionRecording(false);
     }
@@ -82,7 +93,7 @@ public class SampleNativeAppInstallAdMapper extends NativeAppInstallAdMapper {
         mSampleAd.handleClick(view);
     }
 
-    // The Sample SDK doesn't do its own impression/click tracking, instead relies on its
+    // The Sample SDK doesn't do its own impression/click tracking, and instead relies on its
     // publishers calling the recordImpression and handleClick methods on its native ad object. So
     // there's no need to pass it a reference to the View being used to display the native ad. If
     // your mediated network does need a reference to the view, the following method can be used
@@ -103,6 +114,8 @@ public class SampleNativeAppInstallAdMapper extends NativeAppInstallAdMapper {
          * in the publisher’s preferred corner. If a preferred corner is not set, AdChoices
          * should be rendered in the top right corner.
          */
+
+        mSampleAd.registerNativeAdView(view);
 
         if (!(view instanceof ViewGroup)) {
             Log.w(SampleAdapter.TAG, "Failed to show information icon. Ad view not a ViewGroup.");